requirements.txt
Flask==2.1.2
Flask-RESTful==0.3.9
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.37
Flask-JWT-Extended==4.4.1
Flask-Cors==3.0.10
respect-validation==1.3.0
bcrypt==3.2.2
pydantic==1.9.1
requests==2.28.1
paramiko==2.11.0
pexpect==4.8.0
cryptography==37.0.4
psycopg2-binary==2.9.3
GitPython==3.1.27
Jinja2==3.1.2
ua-parser==0.16.1
Werkzeug==2.1.2
supervisor==4.2.4
croniter==1.3.8
python-crontab==2.6.0
gunicorn==20.1.0

python-magic==0.4.27
# important to install libmagic
# here is details
# https://github.com/Yelp/elastalert/issues/1927#issuecomment-689927231